Simplify things around scope tree.
Before, scope tree was just a root node and a map to annotate parse tree node from ANTLR with scope tree node.
But this was implemented in a lazy way and I was stuck with a scope tree root and a map. I was afraid to
forget to add a node to the map or to the tree. So I created ScopeTree<T> which combines a tree and the annotions map within
only one class.

Also remove LCodeParserContext now useless.