Add ParseErrorListener and some tests.
Adding these tests I found an error in the grammar.
Before the parser was accepting entry like:
  a = 123.0;
Interestingly, it was not considered as a valid program but no error was thrown which is not expected at all.
Reason why was how I define a program. It was defined as a list of 0 or more statements.
The cuprite here was to allow 0 statement. The offending input was not considered as a valid statement but
this lead to no errors because 0 statement is allowed. It is potentially a bug in ANTLR.